refactor: create new BlueprintHelper class
