+++ src/Helpers/BlueprintHelper.php
@@ -1,0 +1,38 @@
+<?php
+namespace GravApi\Helpers;
+
+use \Grav\Common\Grav;
+use \Grav\Common\Uri;
+use \Grav\Common\User\Interfaces\UserInterface;
+use \GravApi\Helpers\AuthHelper;
+
+/**
+ * Class BlueprintHelper
+ * @package GravApi\Helpers
+ */
+class BlueprintHelper
+{
+    /**
+     * Returns the API's advanced acess taxonomy field from the user's profile for a given HTTP method.
+     *
+     * This is a hacky workaround to populate the taxonomy field with data from the user's profile.
+     * See https://github.com/getgrav/grav-plugin-admin/issues/1472
+     *
+     * @param string $method One of `get`, `post`, `patch`, `delete`
+     * @return array An empty array will be return if no taxonomy data exists
+     */
+    public static function getUserTaxonomy($method)
+    {
+        $grav = Grav::instance();
+
+        /** @var Uri */
+        $uri = $grav['uri'];
+        // We extract the username from the URL, e.g. /admin/user/tom
+        $username = end($uri->paths());
+
+        /** @var UserInterface */
+        $user = $grav['accounts']->load($username);
+
+        return AuthHelper::getUserTaxonomy($user, $method);
+    }
+}
