+++ src/Helpers/TaxonomyHelper.php
@@ -1,11 +1,6 @@
 <?php
 namespace GravApi\Helpers;
 
-use \Grav\Common\Grav;
-use \Grav\Common\Uri;
-use \Grav\Common\User\Interfaces\UserInterface;
-use \GravApi\Helpers\AuthHelper;
-
 /**
  * Class TaxonomyHelper
  * @package GravApi\Helpers
@@ -66,29 +61,5 @@ class TaxonomyHelper
     public static function hasIntersect($a, $b)
     {
         return count(self::intersect($a, $b)) > 0;
-    }
-
-    /**
-     * Returns the API's advanced acess taxonomy field from the user's profile for a given HTTP method.
-     *
-     * This is a hacky workaround to populate the taxonomy field with data from the user's profile.
-     * See https://github.com/getgrav/grav-plugin-admin/issues/1472
-     *
-     * @param string $method One of `get`, `post`, `patch`, `delete`
-     * @return array An empty array will be return if no taxonomy data exists
-     */
-    public static function getUserTaxonomy($method)
-    {
-        $grav = Grav::instance();
-
-        /** @var Uri */
-        $uri = $grav['uri'];
-        // We extract the username from the URL, e.g. /admin/user/tom
-        $username = end($uri->paths());
-
-        /** @var UserInterface */
-        $user = $grav['accounts']->load($username);
-
-        return AuthHelper::getUserTaxonomy($user, $method);
     }
 }
